Incorporacion de una funcion "Log" para hacer debug, arreglo de problemas de sintaxis y compilacion en el codigo general.

+++ logica/port.h
@@ -19,7 +19,7 @@ public:
   }
 
   std::string GetCode();
-  uint8_t GetAddress();
+  uint32_t GetAddress();
 
   void SetCode(std::string code_);
   void SetAddress(uint8_t address_);
